# booking-service/docs/booking-service-api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Booking Service API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8082/ticketbooking",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/purchase/{purchaseId}": {
      "get": {
        "tags": [
          "Purchase"
        ],
        "summary": "Retrieve a Purchase by Id",
        "description": "Get Purchase order Details by specifying its id. ",
        "operationId": "getTicket",
        "parameters": [
          {
            "name": "purchaseId",
            "in": "path",
            "description": "ID of Purchase order to be retrived",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDTO"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Purchase"
        ],
        "summary": "Cancel a new Ticket",
        "operationId": "cancelTicket",
        "parameters": [
          {
            "name": "purchaseId",
            "in": "path",
            "description": "ID of Purchase order to be canceled",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          },
          "201": {
            "description": "Ticket Purchase order Cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
...
    "/purchase": {
      "post": {
        "tags": [
          "Purchase"
        ],
        "summary": "Purchase a new Ticket",
        "operationId": "purchaseTicket",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Purchase order created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "TicketDTO": {
        "type": "object",
        "properties": {
          "ticketId": {
            "type": "integer",
            "format": "int64"
          },
          "eventId": {
            "type": "integer",
            "format": "int64"
          },
          "ticketType": {
            "type": "string",
            "enum": [
              "NORMAL",
              "VIP"
            ]
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "ON_HOLD",
              "BOOKED"
            ]
          }
        }
      },
...
      "PurchaseRequest": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "string",
            "enum": [
              "NORMAL",
              "VIP"
            ]
          }
        }
      },
...
    }
  }
}